remove unnecessary suffix
Figured out how/why this works:

[TMUX 1.5 patch notes](https://github.com/tmux/tmux/blob/1.5/CHANGES#L33-L34)
These call out using a "tmux;" DCS to pass through escape sequences

"DCS" stands for "Device Control String"

|Name|8-bit Character|7-bit char1|7bit char2|
|:--:|:-------------:|:---------:|:--------:|
|Device Control String|DCS<br />9/0|ESC<br />1/11|P<br />5/0|

Or "\[P". The folks stating that the it is necessary to include a `\[\\` at the end either want a new-line at the end of their output or didn't find this super-buried esoteric patch note.